Move application of dependency substitution into its own class
This allows separating the concern of applying a rule, which is in the new class, from selection based on
the result of the application, which is in `DependencySubstitutionResolver`. This also allows us to reuse
the logic during optional dependency handling code (even if in practice, we don't cache the result of
application on a dependency, so for optional dependencies the rule may be applied at most twice on the
same object).