Avoid computing `ModuleIdentifier` multiple times for the same module
This commit pre-computes a `ModuleIdentifier` from a dependency and passes it around several places where it
is used. This saves a bunch of lookups in the module identifier cache, which tend to be pretty expensive.
Dependency substitution is also passed the original module identifier, in case no substitution is done. This
prevents an additional lookup.