Unpack using a `List` instead of `ImmutableSet.Builder`
This partly undones a previous change where we introduced immutable sets for the sake of comparing faster. However,
the way Guava builds immutable sets doesn't let us tell what the expected size of the set is. As a consequence,
if the number of excludes is >4, then resizes occur, which is quite frequent in the exclude rule merging performance
test.

To avoid this, we collect elements using a regular list, then build an immutable set from this. We use a list
because no wonder what kind of collection you pass in to `copyOf`, it will always call `toArray()` internally. This
means that it's not a good idea to create a set first, then make it immutable because we would hash twice.